requirements.txt
openai>=1.12.0
psycopg2-binary>=2.9.9
pandas>=2.1.4
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0